refactor: extract `formatFiles`
